refactor(shared-server): name public URL ownership directly
